RepoAnalysis/requirements.txt
asgiref==3.7.2
astroid==2.15.5
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==41.0.1
cycler==0.11.0
defusedxml==0.7.1
dill==0.3.6
Django==4.2.1
django-allauth==0.54.0
django-auto-logout==0.5.1
django-environ==0.10.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib==3.4.3
mccabe==0.7.0
numpy==1.24.3
oauthlib==3.2.2
Pillow==9.5.0
platformdirs==3.8.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomlkit==0.11.8
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
